Add files via upload
Made a start on part 3 of the project, created a new class which will search for the trips that have the same times that the user inputs. I have been working on this since 8am and I am not finished yet. I will try finish this class and make a new class which will work on the edges and get the distance of the stops.